created post method for order with comments
